Split hearing into speech and hearing
The package held two subjects that share a word and nothing else. Speech
intelligibility scores a transmission channel and is read by whoever designs a
room or a public address system; hearing conservation is about the ear and is
read by an occupational hygienist. The generated reference and the sidebar had
separated them long ago, which is why the reference had a speech section
drawing from a package called something else; the code is what had not caught
up.

STI, SII and STOI become phonometry.speech. The audiometric thresholds, the
ISO 1999 hearing loss and the ISO 9612 exposure stay in phonometry.hearing,
which is now what its name says. No module is renamed and no reference page
moves, so this is the cheapest of the taxonomy splits and the one that removes
an exception rather than adding a URL.

Compatibility follows the pattern the metrology split established: the three
pre-split module paths still import and delegate with a notice, reading a
moved name from the namespace it left still works, dir() still lists them, and
all of it goes in 5.0. The frozen snapshot of pre-split paths grows by three.

The renderers split the same way, so _plot/hearing.py becomes two modules with
their own translation tables. The speech half gets the Spanish test it never
had, including the four vocal efforts of the ANSI S3.5 spectra, which are
table keys reached through the result rather than written in the renderer.

The snippet gate earned its place here: it caught the SII guide importing the
new speech package into a block that already bound `speech` as the variable
holding the speech spectrum. The variable is renamed.
