+++ site/src/content/docs/es/getting-started.mdx
@@ -154,7 +154,7 @@ float64, así que es seguro pasar directamente la salida de `wavfile.read`.
 
 ## Siguientes pasos
 
-El análisis en octavas de arriba usa el núcleo `filters`, uno de los diecisiete
+El análisis en octavas de arriba usa el núcleo `filters`, uno de los dieciocho
 espacios de nombres de dominio; las secciones de la barra lateral recorren el
 resto, desde la psicoacústica y la acústica de salas, edificación y vibraciones
 hasta el ruido ambiental, aeronáutico y submarino, la electroacústica y la
