+++ site/src/content/docs/es/guides/objective-intelligibility.mdx
@@ -289,4 +289,4 @@ nada implementado más allá de las ecuaciones de 2011/2016 anteriores.
   predice la inteligibilidad a partir de espectros de habla, ruido y umbral.
 - [Bancos de filtros](/phonometry/es/guides/filter-banks/): las bandas de tercio
   de octava en las que el procesado inicial agrupa la DFT.
-- Referencia de la API: [`hearing.objective_intelligibility`](/phonometry/es/reference/api/speech/objective-intelligibility/).
+- Referencia de la API: [`speech.objective_intelligibility`](/phonometry/es/reference/api/speech/objective-intelligibility/).
