+++ site/src/content/docs/es/guides/occupational-exposure.mdx
@@ -266,7 +266,7 @@ ficha en español (*nivel de exposición diario equivalente*, *valor límite de
 exposición*, con coma decimal).
 
 ```python
-from phonometry import hearing, ReportMetadata
+from phonometry import ReportMetadata, hearing
 
 # La jornada de los soldadores del anexo D de la sección 1, con
 # incertidumbre de duración.
