+++ site/src/content/docs/guides/objective-intelligibility.mdx
@@ -285,4 +285,4 @@ beyond the 2011/2016 equations above.
   predicts intelligibility from speech, noise and hearing-threshold spectra.
 - [Filter Banks](/phonometry/guides/filter-banks/): the one-third-octave bands
   the front end groups the DFT into.
-- API reference: [`hearing.objective_intelligibility`](/phonometry/reference/api/speech/objective-intelligibility/).
+- API reference: [`speech.objective_intelligibility`](/phonometry/reference/api/speech/objective-intelligibility/).
