+++ site/src/content/docs/reference/api/speech/sti.md
@@ -1,5 +1,5 @@
 ---
-title: "hearing.sti"
+title: "speech.sti"
 description: "Speech Transmission Index (STI) per IEC 60268-16:2020 (Edition 5)."
 sidebar:
   label: "sti"
