+++ src/phonometry/speech/__init__.py
@@ -1,0 +1,35 @@
+#  Copyright (c) 2026. Jose Manuel Requena Plens
+"""speech domain of phonometry (see module docstrings)."""
+
+from __future__ import annotations
+
+from .objective_intelligibility import STOIResult, stoi
+from .sii import (
+    SII_METHODS,
+    SIIProcedure,
+    SIIResult,
+    StandardSpeechSpectrum,
+    sii_procedure,
+    speech_intelligibility_index,
+    standard_speech_spectra,
+    standard_speech_spectrum,
+)
+from .sti import STIResult, STIWarning, sti_from_impulse_response, stipa, stipa_signal
+
+__all__ = [
+    "SII_METHODS",
+    "SIIProcedure",
+    "SIIResult",
+    "STIResult",
+    "STIWarning",
+    "STOIResult",
+    "StandardSpeechSpectrum",
+    "sii_procedure",
+    "speech_intelligibility_index",
+    "standard_speech_spectra",
+    "standard_speech_spectrum",
+    "sti_from_impulse_response",
+    "stipa",
+    "stipa_signal",
+    "stoi",
+]
